backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.0",
    "fastapi>=0.121.3",
    "langchain>=1.0.8",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.1.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = ["autopep8>=2.3.2", "flake8>=7.3.0", "ruff>=0.14.6"]
